`Cookie`, `CookieCollection` and `HeaderCollection` moved under `yii\http` namespace
