+++ framework/classes.php
@@ -280,15 +280,15 @@ return [
   'yii\web\CompositeUrlRule' => YII2_PATH . '/web/CompositeUrlRule.php',
   'yii\web\ConflictHttpException' => YII2_PATH . '/web/ConflictHttpException.php',
   'yii\web\Controller' => YII2_PATH . '/web/Controller.php',
-  'yii\web\Cookie' => YII2_PATH . '/web/Cookie.php',
-  'yii\web\CookieCollection' => YII2_PATH . '/web/CookieCollection.php',
+  'yii\http\Cookie' => YII2_PATH . '/http/Cookie.php',
+  'yii\http\CookieCollection' => YII2_PATH . '/http/CookieCollection.php',
   'yii\web\DbSession' => YII2_PATH . '/web/DbSession.php',
   'yii\web\ErrorAction' => YII2_PATH . '/web/ErrorAction.php',
   'yii\web\ErrorHandler' => YII2_PATH . '/web/ErrorHandler.php',
   'yii\web\ForbiddenHttpException' => YII2_PATH . '/web/ForbiddenHttpException.php',
   'yii\web\GoneHttpException' => YII2_PATH . '/web/GoneHttpException.php',
   'yii\web\GroupUrlRule' => YII2_PATH . '/web/GroupUrlRule.php',
-  'yii\web\HeaderCollection' => YII2_PATH . '/web/HeaderCollection.php',
+  'yii\http\HeaderCollection' => YII2_PATH . '/http/HeaderCollection.php',
   'yii\web\HtmlResponseFormatter' => YII2_PATH . '/web/HtmlResponseFormatter.php',
   'yii\web\HttpException' => YII2_PATH . '/web/HttpException.php',
   'yii\web\IdentityInterface' => YII2_PATH . '/web/IdentityInterface.php',
