+++ tests/framework/web/UserTest.php
@@ -22,8 +22,8 @@ use Yii;
 use yii\base\Component;
 use yii\base\NotSupportedException;
 use yii\rbac\PhpManager;
-use yii\web\Cookie;
-use yii\web\CookieCollection;
+use yii\http\Cookie;
+use yii\http\CookieCollection;
 use yii\web\ForbiddenHttpException;
 use yii\web\IdentityInterface;
 use yiiunit\TestCase;
